Clean up CSE handling in simplify_inst
Change simplify_inst to use the converter's build API instead of
the CSE API. This is cleaner. This also avoids an issue where we
could simplify inconsistently between src and tgt. The issue is
that the converter has some extra simplifications not done by
simplify_inst, and these were not done in the previous API when
simplify_inst recursively simplified an instruction.