Bring back early exit for identical src/tgt
The check for identical src and tgt was removed in 0c5f4bb because
the way we generate the IR often makes them slightly different,
even if the original GIMPLE was identical.

But there are a few cases where they are identical. This happens
for a few of the nasty GCC test cases, such as gcc.dg/20050321-2.c,
which has a phi node with ~20000 arguments. This is very slow in
smtgcc -- checking the file needs ~30 minutes. But most of the src
and tgt pairs are identical, so bringing back the check for identical
src and tgt makes checking the file finish after ~2 minutes.